requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0
PyPDF2>=3.0.0
pandas>=1.3.0
openpyxl>=3.0.0
python-docx>=0.8.11
Flask>=2.3.0
Werkzeug>=2.3.0
gunicorn>=21.0.0 